Add English responses to the map API requests. Separate the homepage and the menu bar, and allow multiple pages to be added.
