+++ frontend/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600;700&display=swap" rel="stylesheet">
-    <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyBKDDK8WW_VH96hpM7kMWE2wzE2akrFaEo&callback=initMap" async defer></script>
+    <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyBKDDK8WW_VH96hpM7kMWE2wzE2akrFaEo&callback=initMap&language=en" async defer></script>
 
     <title>SCUVMS</title>
   </head>
